+++ packages/sdk/examples/next/pages/index.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2022 AccelByte Inc. All Rights Reserved
+ * Copyright (c) 2018-2023 AccelByte Inc. All Rights Reserved
  * This is licensed software from AccelByte Inc, for limitations
  * and restrictions contact your company contract manager.
  */
@@ -87,7 +87,7 @@ export default function Home({ data }: InferGetServerSidePropsType<typeof getSer
           </div>
         </div>
 
-        <div className={styles.responses}>
+        <div className={styles.responses} id="response">
           {Object.keys(data).map(key => (
             <Collapsible key={key} json={data[key as keyof TestSdkReturnType]} title={key} />
           ))}
@@ -100,10 +100,10 @@ export default function Home({ data }: InferGetServerSidePropsType<typeof getSer
 // composing components.
 function Collapsible(props: { title: string; json: any }) {
   return (
-    <details>
+    <details id={'toogle' + props.title}>
       <summary>{props.title}</summary>
 
-      <pre>{JSON.stringify(props.json, null, 2)}</pre>
+      <pre id={props.title}>{JSON.stringify(props.json, null, 2)}</pre>
     </details>
   )
 }